api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
psycopg2-binary>=2.9.9
anthropic>=0.39.0
pyyaml>=6.0
slowapi>=0.1.9
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.24.0
